feat: Major app improvements - GST reports, UX enhancements, and new features
- GST Reports: GSTR-1 review screen, Excel/PDF export, HSN-category mapping
- Billing: Receipt PDF generation, UPI payment integration, loose items support
- Inventory: Product image support, improved import/parsing
- UI/UX: Privacy overlay for sensitive data, vendor/customer detail sheets
- Sync: Improved batch sync, outbox codec, immediate sync manager
- Theme: Updated colors, launcher icons, app branding to thisizbusiness
- Added new utilities: Receipt rendering, UPI intent helper, product image store
- Database: Schema updates for GST invoicing support
- Supabase: Updated sync functions and SQL migrations

+++ app/src/main/java/com/kiranaflow/app/data/local/GstReportRows.kt
@@ -41,7 +41,7 @@ data class GstSaleLineRow(
     val itemNameSnapshot: String,
 
     @ColumnInfo(name = "qty")
-    val qty: Int,
+    val qty: Double,
 
     @ColumnInfo(name = "unit")
     val unit: String,
@@ -73,6 +73,8 @@ data class GstSaleLineRow(
     @ColumnInfo(name = "itemGstPercentage")
     val itemGstPercentage: Double?
 )
+
+
